+++ src/ui/__init__.py
@@ -18,6 +18,7 @@ from .entertainment import register_entertainment           # noqa: F401
 from .deployos import register_deployos                    # noqa: F401
 from .pxe import register_pxe                              # noqa: F401
 from .tiendo import register_tiendo                        # noqa: F401
+from .gopy import register_gopy                            # noqa: F401
 
 
 def register_all(app):
@@ -40,4 +41,5 @@ def register_all(app):
     register_deployos(app)
     register_pxe(app)
     register_tiendo(app)
+    register_gopy(app)
     return app